task/requirements.txt
Flask==3.1.0
redis==5.2.1
python-dotenv==1.1.0